Fix a bug with handling the completion type in a hidden friend.
The issue was that with the init function being a non-template friend
function defined in a class template, it gets instantiated more eagerly
than member functions. This caused a bug in the case of non-block scoped
cuda::barrier with a non-default-constructible completion function fail
to compile when the class itself is instantiated, because the default
parameter to `init` that was being used was eagerly instantiated, and
that failed (since the default value attempted to default construct a
completion function).

Fixing this by splitting the friend into two overloads, one with a
completion function argument and one without. (Note that having a
*constructor* with the exact signature is *fine*, because it is not as
eagerly instantiated as the friend; it only gets instantiated on use,
while all friend functions are instanted immediately, because they
generate real functions, not templates.)